+++ src/release_validation.py
@@ -1,0 +1,3 @@
+"""Wrapper publico para a validacao oficial de release."""
+
+from .interfaces.release_validation import *  # noqa: F401,F403